Fix django auth login redirection
The django authentication decorator login_required, redirects to
settings.LOGIN_URL. If it is missing, it redirects to /accounts/login.

We did not have the setting enable, not a proper page where to
redirect. This cause users not logged in to see a 404 error when
accessing courseware directly, by using a link or a bookmark.

[#41499213]