Modify course registration flow for new login and registration pages
Now that we are using separate pages for login and registration rather than
modals, clicking the registration button for a course should direct an
unauthenticated user to the registration page, and the user should be enrolled
in the course upon successful registration. Likewise, an unauthenticated user
attempting to unenroll from a course should be directed to the login page and
subsequently unenrolled from the course upon successful login. The enrollment
change service also now uses HTTP status codes rather than JSON to communicate
status to the front end.