Added functionality to bypass alerts
This is done with the following steps:
'I confirm all alerts' means that all alert and confirm windows are returned and returned true respectively
'I dismiss all alerts' means that all confirm windows are returned false
'I answer all prompts with "([^"]*)"' means that all prompts are returned with the given string

Please note that these settings are on a PER PAGE basis.  This means that for best results, the step must be given right before the alert is generated.