Removing JavaScript qTip tooltips.
Because sometimes the tooltips generated by qTip get in the way, and the controls becomes
not responsive, it was decided to use standard title attributes for tips.

Also along the way, an error was discoevered in Jasmine tests. It was fixed, but that test is
failing so it was marked specifically to be skipped when all VideoAlpha tests run.