Move loading of YouTube API to video module.
The availability of the YouTube API will be stored in student's settings.
The YouTube API is loaded asynchronously now, so no need to pass a
parameter to the front-end telling it if YouTube API should not be loaded
because of it's unavailability.

Removing loading of YouTube API from Studio RequireJS config. Now loading
of YouTube API is handled by Video module.

BLD-531.