Fix a url processing error in 02_html5_video.js
As the code in this .js will break the url's integrity when there is already a query string inside the video source's url and cause the url to be invalid in some cases (for example, when working with a url from Windows Azure's Media Service, the appended '?' will cause the url being invalid).
I modified the code by first checking whether the url has already had a query string, and if so, then use '&' instead of '?' to prevent the breaking, and it works for the url from Windows Azure's Media Service.