i18n: Fix the chinese name of chinese language
The current version of chinese language's chinese name is not political neutral.
In fact, which form of chinese (traditional or simplified) can be called "正體"(in traditional chinese) or "正体"(in simplified chinese) is not globally defined.
So, I change the name according to the word used by Microsoft Windows, which is more neutral.