Translation fix in studio: For better translation
As the sentence in unit.html is splitted into three parts due to the varible release_date may be None, it breaks the integrity of a sentence.
Different language may have different word or event grammer order, so if a sentence is breaked into three parts, translators cannot adjust the word orders to create a good translation.
So this fix tries to combine the sentence together.