Separate edX / Open edX footer.
We need to use an Open edX footer by default (and that's what all non-edx.org
instances should base their footer off) and only use the edX footer on
edx.org.

This commit uses `SITE_NAME` to determine which footer to use.

Sites using theming or micro-sites with a `SITE_NAME` not ending in `edx.org`
should not be affected.