User registration prevents using password as username.
Some users erroneously set their password as their username with the original layout, because the username field directly followed the password field. Users may be accustomed to the common occurrence of a password confirmation field directly following the password field.
To fix the issue, I did the following:
   - Moved the existing username and real name form fields above the password field.
   - Added a validation in the create_account handler in common/djangoapps/student/views.py, which confirms that the password field does not match the username field.

New tests created to check the added functionality.