Instructors can view previously sent email content
Previously on the send email page of the instructor dashboard, instructors could only view
task information about emails they've sent for their course in the past.
In addition to this, I've now added the ability to see the content of all previously sent emails.

A "Sent Email History" button has been added to the page. When clicked, a table displaying the
subject line, number of emails sent, and date/time of submission for each previously sent email
is created. An instructor can then click on any subject line to see the content of that email,
displayed in a modal window that appears on the page.

The window is also equipped with a "copy email to editor" button, which copies the emails contents
to the tinyMCE editor, so that an instructor can easily resend an email that they've sent
in the past.