Bugfix: Initialize new tabname to "Empty"
In Studio, when the course administrator creates a new tab, its name was being
initialized as null in the mongo database. This caused students in LMS to get a
server error when trying to access the course page, since LMS was trying to
process the null tabname for translation, etc. Now tabname is initialized as
'Empty', which is how it was being done in the frontend.

STUD-2000