Fix false positive test in cms acceptance
This test was passing in Jenkins, but failing in our Solano runs (and locally on firefox).
This is because it was attaching a file that was not there. This change fixes the test
and makes the attachment method smarter so it will fail if it tries attaching a non-existent
file.