Add HTTP_RANGE compatibility for ContentStore file streaming
Currently, users can only download ContentStore files from first byte to
last byte.

With this change, when a request to the ContentStore includes the HTTP
"Range" parameter, it is parsed and StaticContentStream will stream the
requested bytes.

This change makes possible to stream video files (.mp4 especially) from
the ContentStore. Users can now seek a specific time in the video
without loading all the file.

This is useful for courses with a limited number of students that
doesn't require a dedicated video server.