Create shard 3 for bok-choy; handle unknown shards.
This accomplishes three things:
1. Create a 3rd bok-choy shard
2. Allow the script to run 'all' shards
3. If the script receives an unknown shard, it will pass through empty tests.

The last one (3) is intended to allow us to expand to additional shards
on our Jenkins infrastructure. If we have, for example, a 4th shard introduced,
but someone attemps to run a commit that is older (and therefore still
has 3 shards); the script will just pass-through shard '4'. This means
older commits can still get tested. In that case, the older commit
will still test all cases, because the last shard (n) is basically an
'everything else' shard.