Remove antiquated cleanup script
This script was originally written to clean up XML scripts. In the more
than two years since it was originally written, the only updates it has
received have been to remove PEP8/PyLint/quality violations.

It's hard to imagine that this is still useful to anyone, though I'm
open to discussion.