Vote buttons should be disabled while in use
Clicking quickly on up/down votes in forums generates a race condition
where the votes can continue going up or down (and I can get e.g. large
negative numbers). This is client-side; the server-side is not affected.

This is caused by safeAjax calling code (including updateWithUndo) not
handling the possible undefined return value. This is returned in the case
where the element that triggers the event is already disabled.

Corrected typo in DiscussionThreadView cleanup: @responseRequest -> @responsesRequest

Resolves TNL-1061