Lms should be independent of mathjax
When cdn is down specially in the country like china
and others user face problem as application doesn't
work as intended. Instead of making mathjax hard
dependecy I have converted it into soft dependey
so that it won't cause user problem to use application
smothly if mathjax is not present.

TNL-1498