Support filtering team membership
Adds tests and functionality to support filtering the results returned
from team_membership endpoint with optional course_id parameter. Will
return 400 Bad Request in the event that the given course_id does not
match up with the provided team_id (team.course_id does not match), or
if the user is not enrolled in the course given.

Documentation has been added to the relevant API page on the wiki:
https://openedx.atlassian.net/wiki/display/TNL/Team+API