Allow user partition version >= current version.
If code is deployed that updates the user partition version,
then the code is rolled back, we may see user partition versions
greater than the currently deployed version.

The previous behavior was to raise a TypeError when this occurred;
the new behavior assumes that the newer version is backwards
compatible and tries to load the partition.