Potential fix for flaky TeamDiscussionView js tests
Adding waitsFor check in the beforeEach method. The condition I'm
waiting for here is shamelessly borrowed from bok-choy's
wait_for_ajax._is_ajax_finished() method:
https://github.com/edx/bok-choy/blob/master/bok_choy/page_object.py#L431

This seems to fix the errors I was able to generate with the
set_timeout, which themselves looked ver similar to the failing
master tests, so I *think* this will fix our woes, but since I was
never able to get the tests to fail exactly the same way I can't be sure.

Since there don't seem to be any drawbacks, this is going into mainline
as-is and we'll monitor for the sporadic failures over the next few
weeks.