Adjust the i18n extraction rules to get all the strings we should get.
There's a move afoot to put assets inside individual Django apps,
instead of in upper-level "templates" directories.  These i18n
configuration files have to take this into account.

.underscore files are easy to find by their unique file extension, so
the rule can be simple.  Mako files are .html, so we have to be more
nuanced about it.  We don't want to get pure HTML files, and we don't
want to get .html files for tests, even if they are Mako templates.